requirements.txt

numpy>=1.19.2
scipy>=1.6.1
tqdm>=4.56.0
matplotlib>=3.3.4
h5py>=2.10.0
pandas>=1.2.2
seaborn>=0.11.1
pyproj>=2.6.1.post1
torchsummary>=1.5.1
scikit-learn>=0.24.1
xgboost>=1.3.3